requirements.txt
holidays==0.13
matplotlib==3.1.3
numpy==1.18.1
pandas==1.2.4
pandas_market_calendars==3.3
python_dateutil==2.8.2
scikit_learn==1.1.1
scipy==1.4.1
statsmodels==0.11.0
wrds==3.1.1
